// src/app/enums/enums.ts
/* tslint:disable */
/* eslint-disable */
// Generated using typescript-generator version 2.24.612 on 2025-02-17 14:49:41.

export enum ReportType {
    DAMAGE = "DAMAGE",
    LATE_RETURN = "LATE_RETURN",
    LOST_KEY = "LOST_KEY",
    OTHER = "OTHER",
}

export enum ReservationStatus {
    PENDING = "PENDING",
    CONFIRMED = "CONFIRMED",
    CANCELLED = "CANCELLED",
    COMPLETED = "COMPLETED",
}

export enum VehicleStatus {
    AVAILABLE = "AVAILABLE",
    UNDER_MAINTENANCE = "UNDER_MAINTENANCE",
    INSURANCE_EXPIRED = "INSURANCE_EXPIRED",
    TECHNICAL_CONTROL_EXPIRED = "TECHNICAL_CONTROL_EXPIRED",
}